[feat] Add main script for getting data ready
