+++ utils/__init__.py
@@ -1,0 +1,1 @@
+from .prepare_data import main as run_prepare_data
